Iris pint cf_units compatibility layer (#7027)
* Separate the cf_units emulation layer from permanent behaviours of the Iris 'PintUnit'.

* Move all units code into a separate iris.common.units submodule.

* Knockon test fixes.

* Extend special cfpint installs to doctests, wheels and RTD build.

* Adjust docstrings, add deprecations for all cf_units compatibility methods.

* Use enums and constants for pint unit categories.

* Update lib/iris/experimental/units.py


* Update lib/iris/experimental/units.py


* Update lib/iris/experimental/units.py


* Update lib/iris/common/units/_pint.py


* Update lib/iris/common/units/_pint.py


* Update lib/iris/common/units/_pint.py


* Fix deprecation versions.

* Small review actions.

* Use the category constants.

* Some suggestions for fully harnessing Enums. (#126)

---------
