+++ lib/iris/tests/unit/common/units/test_make_unit.py
@@ -8,7 +8,7 @@ from cf_units import Unit as cf_unit
 import numpy as np
 from pint import Unit as pint_unit
 
-from iris.common.mixin import CfpintUnit, CfUnit, make_unit
+from iris.common.units import CfUnit, PintUnit, make_unit
 from iris.experimental.units import USE_CFPINT
 
 
@@ -17,8 +17,8 @@ def test_from_pint():
     no_context_unit = make_unit(unit)
     with USE_CFPINT.context():
         context_unit = make_unit(unit)
-    assert isinstance(no_context_unit, CfpintUnit)
-    assert isinstance(context_unit, CfpintUnit)
+    assert isinstance(no_context_unit, PintUnit)
+    assert isinstance(context_unit, PintUnit)
 
 
 def test_from_cf_units():
@@ -41,6 +41,6 @@ def test_from_number():
         int_unit = make_unit(1)
         float_unit = make_unit(1.0)
         numpy_unit = make_unit(np.float64(1))
-        assert isinstance(int_unit, CfpintUnit)
-        assert isinstance(float_unit, CfpintUnit)
-        assert isinstance(numpy_unit, CfpintUnit)
+        assert isinstance(int_unit, PintUnit)
+        assert isinstance(float_unit, PintUnit)
+        assert isinstance(numpy_unit, PintUnit)
