+++ lib/iris/tests/unit/common/units/test_pintunits.py
@@ -14,32 +14,31 @@ import numpy as np
 import pytest
 
 import iris.common.mixin
-from iris.common.mixin import CfpintUnit
+from iris.common.units import PintUnit
 
 
 def test_num2date():
-    unit = CfpintUnit("days since 1970-01-01")
+    unit = PintUnit("days since 1970-01-01")
     vals = np.array([1.0, 2])
     result = unit.num2date(vals)
     assert np.all(result == [datetime(1970, 1, 2), datetime(1970, 1, 3)])
 
 
 def test_date2num():
-    unit = CfpintUnit("days since 1970-01-01")
+    unit = PintUnit("days since 1970-01-01")
     vals = np.array([datetime(1970, 1, 2), datetime(1970, 1, 3)])
     result = unit.date2num(vals)
     assert np.all(result == [1.0, 2])
 
 
 def test_nounit_eq():
-    unit = CfpintUnit("m")
+    unit = PintUnit("m")
     assert unit != "no_unit"
 
 
 def test_calendar():
-    unit = CfpintUnit("days since 1970-01-01", calendar="360_day")
+    unit = PintUnit("days since 1970-01-01", calendar="360_day")
     # NOTE: no <>, due to "backwards compatibility" for assert_CDL
-    # TODO: remove the CfpintUnit._REPR_NO_LTGT
     assert repr(unit) == "Unit('days since 1970-01-01', calendar='360_day')"
     # TODO: should really add the calendar to the string format
     #   I think this is a bit horrible,
@@ -48,56 +47,56 @@ def test_calendar():
     assert str(unit) == "days since 1970-01-01"
 
 
-_UNKNOWN_NAMES = iris.common.mixin.CfpintUnit._IRIS_EXTRA_CATEGORIES["unknown"]
-_NOUNIT_NAMES = iris.common.mixin.CfpintUnit._IRIS_EXTRA_CATEGORIES["no_unit"]
+_UNKNOWN_NAMES = PintUnit._IRIS_CATEGORY_ALIASES[PintUnit.UnitCategory.unknown]
+_NOUNIT_NAMES = PintUnit._IRIS_CATEGORY_ALIASES[PintUnit.UnitCategory.no_unit]
 
 
 class TestFromUnit:
-    """Test CfpintUnit creation from various sources."""
+    """Test PintUnit creation from various sources."""
 
     def test_none_unknown(self):
-        unit = CfpintUnit.from_unit(None)
+        unit = PintUnit.from_unit(None)
         assert unit.category == "unknown"
         assert unit.calendar is None
         assert unit == "unknown"
 
     @pytest.mark.parametrize("name", _UNKNOWN_NAMES)
     def test_str_unknown(self, name):
-        unit = CfpintUnit.from_unit(None)
+        unit = PintUnit.from_unit(None)
         assert unit.category == "unknown"
         assert unit.calendar is None
         assert all(unit == form for form in _UNKNOWN_NAMES)  # string equivalence
 
     def test_cfunits_unknown(self):
         cfunit = cf_units.Unit(None)
-        unit = CfpintUnit.from_unit(None)
+        unit = PintUnit.from_unit(None)
         assert unit.is_unknown()
 
     @pytest.mark.parametrize("name", _NOUNIT_NAMES)
     def test_str_nounit(self, name):
-        unit = CfpintUnit.from_unit(name)
+        unit = PintUnit.from_unit(name)
         assert unit.category == "no_unit"
         assert unit.calendar is None
         assert all(unit == form for form in _NOUNIT_NAMES)  # string equivalence
 
     def test_cfunits_nounit(self):
         cfunit = cf_units.Unit("no_unit")
-        unit = CfpintUnit.from_unit(cfunit)
+        unit = PintUnit.from_unit(cfunit)
         assert unit.is_no_unit()
 
     def test_str(self):
-        unit = CfpintUnit.from_unit("m")
+        unit = PintUnit.from_unit("m")
         assert unit == "metres"  # string equivalence
         assert unit.calendar is None
         assert unit.category == "regular"
 
     def test_cfunits(self):
         cfunit = cf_units.Unit("m")
-        unit = CfpintUnit.from_unit(cfunit)
+        unit = PintUnit.from_unit(cfunit)
         assert unit == "metre"
 
     def test_str_date(self):
-        unit = CfpintUnit.from_unit("days since 1970-01-01")
+        unit = PintUnit.from_unit("days since 1970-01-01")
         assert unit == "days since 1970-01-01"
         assert unit.category == "regular"
         assert unit.is_datelike()
@@ -105,7 +104,7 @@ class TestFromUnit:
 
     @pytest.mark.skip("from_unit does not support calendar (yet?)")
     def test_str_date_calendar(self):
-        unit = CfpintUnit.from_unit("days since 1970-01-01", calendar="360_day")
+        unit = PintUnit.from_unit("days since 1970-01-01", calendar="360_day")
         # YUCK!! cf_units compatibility
         # TODO: this needs to change
         assert unit == "days since 1970-01-01"
@@ -115,7 +114,7 @@ class TestFromUnit:
 
     def test_cfunits_date(self):
         cfunit = cf_units.Unit("hours since 1800-03-09 11:11")
-        unit = CfpintUnit.from_unit(cfunit)
+        unit = PintUnit.from_unit(cfunit)
         # NB time ref is reproduced as-is
         # TODO: should get normalised
         assert unit == "hours since 1800-03-09 11:11"
@@ -124,7 +123,7 @@ class TestFromUnit:
 
     def test_cfunits_date_calendar(self):
         cfunit = cf_units.Unit("hours since 1800-03-09 11:11", calendar="365_day")
-        unit = CfpintUnit.from_unit(cfunit)
+        unit = PintUnit.from_unit(cfunit)
         # NB time ref is reproduced as-is
         # TODO: should get normalised
         assert unit == "hours since 1800-03-09 11:11"
